+++ playgrounds/demo/app/pages/components/file-upload.vue
@@ -3,12 +3,12 @@ import * as z from 'zod'
 import type { FormSubmitEvent } from '@bitrix24/b24ui-nuxt'
 import theme from '#build/b24ui/file-upload'
 import usePageMeta from './../../composables/usePageMeta'
-import ExampleGrid from '../../components/ExampleGrid.vue'
-import ExampleCard from '../../components/ExampleCard.vue'
 import ImageIcon from '@bitrix24/b24icons-vue/outline/ImageIcon'
 import UploadIcon from '@bitrix24/b24icons-vue/outline/UploadIcon'
 
 usePageMeta.setPageTitle('FileUpload')
+
+const isUseBg = ref(true)
 
 const colors = Object.keys(theme.variants.color) as Array<keyof typeof theme.variants.color>
 const sizes = Object.keys(theme.variants.size) as Array<keyof typeof theme.variants.size>
@@ -21,6 +21,7 @@ const size = ref<keyof typeof theme.variants.size>('md')
 const variant = ref<keyof typeof theme.variants.variant>('area')
 const layout = ref<keyof typeof theme.variants.layout>('grid')
 const position = ref<keyof typeof theme.variants.position>('outside')
+const preview = ref(true)
 
 const MAX_FILE_SIZE = 2 * 1024 * 1024 // 2MB
 const MIN_DIMENSIONS = { width: 200, height: 200 }
@@ -90,30 +91,34 @@ async function onSubmit(event: FormSubmitEvent<schema>) {
 </script>
 
 <template>
-  <ExampleGrid v-once class="mb-2">
-    <ExampleCard title="options">
-      <B24Separator class="mb-4" />
-      <div class="flex flex-wrap items-center gap-3">
-        <B24FormField label="color" class="w-full">
-          <B24Select v-model="color" :items="colors" />
-        </B24FormField>
-        <B24FormField label="size" class="w-full">
-          <B24Select v-model="size" :items="sizes" />
-        </B24FormField>
-        <B24FormField label="variant" class="w-full">
-          <B24Select v-model="variant" :items="variants" />
-        </B24FormField>
-        <B24FormField label="layout" class="w-full">
-          <B24Select v-model="layout" :items="layouts" />
-        </B24FormField>
-        <B24FormField label="position" class="w-full">
-          <B24Select v-model="position" :items="positions" />
-        </B24FormField>
-      </div>
-    </ExampleCard>
+  <B24PageGrid v-once class="lg:grid-cols-4 gap-5">
+    <B24Card variant="outline">
+      <template #header>
+        <div class="flex flex-row items-center justify-between gap-2">
+          <ProseH5 class="mb-0">
+            Options
+          </ProseH5>
+          <B24Switch v-model="isUseBg" label="isUseBg" size="xs" />
+        </div>
+      </template>
 
-    <ExampleCard title="ver1" class="md:col-span-1">
-      <B24Separator class="mb-4" />
+      <div class="mb-4 flex flex-col gap-4">
+        <B24Select v-model="color" :items="colors" />
+        <B24Select v-model="size" :items="sizes" />
+        <B24Select v-model="variant" :items="variants" />
+        <B24Select v-model="layout" :items="layouts" />
+        <B24Select v-model="position" :items="positions" />
+        <B24Switch v-model="preview" label="preview" size="xs" />
+      </div>
+    </B24Card>
+
+    <B24Card :variant="isUseBg ? 'outline-no-accent' : 'plain-no-accent'" class="md:col-span-1">
+      <template #header>
+        <ProseH5 class="mb-0">
+          ver1
+        </ProseH5>
+      </template>
+
       <div class="flex flex-col items-center gap-8">
         <B24Form :schema="schema" :state="state" class="space-y-4" @submit="onSubmit">
           <B24FormField
@@ -123,7 +128,7 @@ async function onSubmit(event: FormSubmitEvent<schema>) {
             :color="color"
             :size="size"
           >
-            <B24FileUpload v-slot="{ open, removeFile }" v-model="state.avatar" accept="image/*">
+            <B24FileUpload v-slot="{ open, removeFile }" v-model="state.avatar" accept="image/*" :preview="preview">
               <div class="flex flex-wrap items-center gap-3">
                 <B24Avatar size="lg" :src="state.avatar ? createObjectUrl(state.avatar) : undefined" :icon="ImageIcon" />
 
@@ -147,10 +152,15 @@ async function onSubmit(event: FormSubmitEvent<schema>) {
           <B24Button label="Submit" type="submit" />
         </B24Form>
       </div>
-    </ExampleCard>
+    </B24Card>
 
-    <ExampleCard title="ver2" class="md:col-span-2">
-      <B24Separator class="mb-4" />
+    <B24Card :variant="isUseBg ? 'outline-no-accent' : 'plain-no-accent'" class="md:col-span-2">
+      <template #header>
+        <ProseH5 class="mb-0">
+          ver2
+        </ProseH5>
+      </template>
+
       <div class="flex flex-col items-center gap-8">
         <B24FileUpload
           v-model="value"
@@ -159,15 +169,21 @@ async function onSubmit(event: FormSubmitEvent<schema>) {
           :variant="variant"
           :layout="layout"
           :position="position"
+          :preview="preview"
           label="Drop your image here"
           description="SVG, PNG, JPG or GIF (max. 2MB)"
           :class="variant === 'area' ? 'w-full min-h-44' : ''"
         />
       </div>
-    </ExampleCard>
+    </B24Card>
 
-    <ExampleCard title="ver3" class="md:col-span-4">
-      <B24Separator class="mb-4" />
+    <B24Card :variant="isUseBg ? 'outline-no-accent' : 'plain-no-accent'" class="md:col-span-4">
+      <template #header>
+        <ProseH5 class="mb-0">
+          ver3
+        </ProseH5>
+      </template>
+
       <div class="flex flex-col items-center gap-8">
         <B24FileUpload
           v-model="valueMultiple"
@@ -176,6 +192,7 @@ async function onSubmit(event: FormSubmitEvent<schema>) {
           :variant="variant"
           :layout="layout"
           :position="position"
+          :preview="preview"
           :icon="ImageIcon"
           label="Drop your images here"
           description="SVG, PNG, JPG or GIF (max. 2MB)"
@@ -221,6 +238,6 @@ async function onSubmit(event: FormSubmitEvent<schema>) {
           </template>
         </B24FileUpload>
       </div>
-    </ExampleCard>
-  </ExampleGrid>
+    </B24Card>
+  </B24PageGrid>
 </template>