+++ test/components/FileUpload.spec.ts
@@ -65,6 +65,7 @@ describe('FileUpload', () => {
     ['with multiple', { props: { ...props, multiple: true } }],
     ['without dropzone', { props: { dropzone: false } }],
     ['without interactive', { props: { interactive: false } }],
+    ['without preview', { props: { ...props, preview: false } }],
     ['with required', { props: { required: true } }],
     ['with disabled', { props: { disabled: true } }],
     ['with fileIcon', { props: { ...props, fileIcon: Cross30Icon } }],