+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## [2.1.2](https://github.com/bitrix24/b24ui/compare/v2.1.1...v2.1.2) (2025-11-xx)
 
+* feat(FileUpload): add `preview` prop
 * fix(Link): partial extend for `vue-router` and `inertia`
 * fix(ProseCallout): add MdnWebDocIcon|InfoCircleIcon
+* docs(config): add extraAllowedHosts
 
 ## [2.1.1](https://github.com/bitrix24/b24ui/compare/v2.1.0...v2.1.1) (2025-11-11)